requirements.txt
# Core
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
cssutils>=2.7.0

# Networking & DNS
aiodns>=3.0.0
cryptography>=41.0.0
pyOpenSSL>=23.0.0
dnspython>=2.4.0
whois>=0.9.0
tldextract>=5.0.0

# Performance
psutil>=5.9.0
orjson>=3.9.0
ujson>=5.8.0

# Analysis
nltk>=3.8.0
textstat>=0.7.0
yake>=0.4.8
syllables>=0.1.0
langdetect>=1.0.9

# Reporting
jinja2>=3.1.0
markdown>=3.5.0
matplotlib>=3.7.0
plotly>=5.17.0
wordcloud>=1.9.0
pdfkit>=1.0.0
weasyprint>=60.0
reportlab>=4.0.0

# CLI & UI
click>=8.1.0
rich>=13.0.0
colorama>=0.4.6
tabulate>=0.9.0

# Utilities
python-dateutil>=2.8.2
urlextract>=1.8.0
pandas>=2.0.0
numpy>=1.24.0
redis>=5.0.0  # For caching

# Additional scanner dependencies
dnspython>=2.4.0
cryptography>=41.0.0
pyOpenSSL>=23.0.0
whois>=0.9.0
